Update static pages and improve navigation layout
This commit updates the display content and routing for the static pages in the application as follows: 'home', 'about', 'projects', 'contact', and 'resume'. The home page now shows the current time. The underlying HTML for the static pages was removed, and the correct links for navigation have been applied. Improved CSS styling has also been added to application.css to better present the content layout.